// src/features/bankDetailsDistributor/bank_search_branch/types.ts
export enum ACTION {
  ON_SELECT_IFSC = "ON_SELECT_IFSC",
  CLEAR_SEARCH = "CLEAR_SEARCH",
  SEARCH_IFSC_ACTION = "SEARCH_IFSC_ACTION",
  GO_BACK = "GO_BACK",
}
export type IFSCCodePayload = { ifscCode: string; bankAccountNumber: string };
export type IFSCSearchActionPayload = {
  bankCode: string;
  value: string;
  bankName: string;
  bankAccountNumber: string
};